split out random functionality into seperate component
